Switched IDE's cuz I couldn't be asked with Studio's bum ass commit problems.
But in terms of the code I started with  Round 1 configuration for both the confereces, I probably will have everything be run in one function except the automatic code cuz it will be hard to not have a "globaly variable". Something I would understnad only. But yeah whatever.

This whole file is for me to draft how the basic strucutre for the code that I will then rewrite using a CSV database and Queue algorithims